@@ -1,25 +1,25 @@
 name: eyflutter_uikit
 description: eyflutter项目工具库
-version: 0.0.1+3
+version: 0.0.1+5
 homepage: https://github.com/eyinfo/eyflutter_uikit
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
   uni_links: ^0.5.1
-  eyflutter_core: ^0.0.1+12
-  flutter_easyrefresh: ^2.1.8
+  eyflutter_core: ^0.0.3+1
+  flutter_easyrefresh: ^2.2.1
   flutter_gifimage: ^1.0.1
   convex_bottom_bar: ^2.7.1+2
   percent_indicator: ^3.0.1
   meta: ^1.3.0
   extended_text: ^7.0.0
-  cached_network_image: ^3.1.0
-  flutter_cache_manager: ^3.1.2
+  cached_network_image: ^3.2.0
+  flutter_cache_manager: ^3.3.0
   http: ^0.13.3
 
 dev_dependencies:
