@@ -5,26 +5,26 @@ import 'flutter_custom_dialog_widget.dart';
 class YYDialog {
   //================================弹窗属性======================================
   List<Widget> widgetList = []; //弹窗内部所有组件
-  static BuildContext _context; //弹窗上下文
-  BuildContext context; //弹窗上下文
+  static BuildContext? _context; //弹窗上下文
+  BuildContext? context; //弹窗上下文
 
-  double width; //弹窗宽度
-  double height; //弹窗高度
+  double? width; //弹窗宽度
+  double? height; //弹窗高度
   Duration duration = Duration(milliseconds: 250); //弹窗动画出现的时间
   Gravity gravity = Gravity.center; //弹窗出现的位置
   bool gravityAnimationEnable = false; //弹窗出现的位置带有的默认动画是否可用
   Color barrierColor = Colors.black.withOpacity(.3); //弹窗外的背景色
-  BoxConstraints constraints; //弹窗约束
-  Function(Widget child, Animation<double> animation) animatedFunc; //弹窗出现的动画
+  BoxConstraints? constraints; //弹窗约束
+  Function(Widget child, Animation<double> animation)? animatedFunc; //弹窗出现的动画
   bool barrierDismissible = true; //是否点击弹出外部消失
   EdgeInsets margin = EdgeInsets.all(0.0); //弹窗布局的外边距
 
-  Decoration decoration; //弹窗内的装饰，与backgroundColor和borderRadius互斥
+  Decoration? decoration; //弹窗内的装饰，与backgroundColor和borderRadius互斥
   Color backgroundColor = Colors.white; //弹窗内的背景色
   double borderRadius = 0.0; //弹窗圆角
 
-  Function() showCallBack; //展示的回调
-  Function() dismissCallBack; //消失的回调
+  Function()? showCallBack; //展示的回调
+  Function()? dismissCallBack; //消失的回调
 
   get isShowing => _isShowing; //当前 弹窗是否可见
   bool _isShowing = false;
@@ -34,7 +34,7 @@ class YYDialog {
     _context = ctx;
   }
 
-  YYDialog build([BuildContext ctx]) {
+  YYDialog build([BuildContext? ctx]) {
     if (ctx == null && _context != null) {
       this.context = _context;
       return this;
@@ -94,7 +94,7 @@ class YYDialog {
     fontSize1,
     fontWeight1,
     fontFamily1,
-    VoidCallback onTap1,
+    VoidCallback? onTap1,
     text2,
     color2,
     fontSize2,
@@ -155,10 +155,10 @@ class YYDialog {
   }
 
   YYDialog listViewOfListTile({
-    List<ListTileItem> items,
-    double height,
+    List<ListTileItem>? items,
+    double? height,
     isClickAutoDismiss = true,
-    Function(int) onClickItemListener,
+    Function(int)? onClickItemListener,
   }) {
     return this.widget(
       Container(
@@ -166,7 +166,7 @@ class YYDialog {
         child: ListView.builder(
           padding: EdgeInsets.all(0.0),
           shrinkWrap: true,
-          itemCount: items.length,
+          itemCount: items?.length ?? 0,
           itemBuilder: (BuildContext context, int index) {
             return Material(
               color: Colors.white,
@@ -180,15 +180,15 @@ class YYDialog {
                       dismiss();
                     }
                   },
-                  contentPadding: items[index].padding ?? EdgeInsets.all(0.0),
-                  leading: items[index].leading,
+                  contentPadding: items?[index].padding,
+                  leading: items?[index].leading,
                   title: Text(
-                    items[index].text ?? "",
+                    items?[index].text ?? "",
                     style: TextStyle(
-                      color: items[index].color ?? null,
-                      fontSize: items[index].fontSize ?? null,
-                      fontWeight: items[index].fontWeight,
-                      fontFamily: items[index].fontFamily,
+                      color: items?[index].color,
+                      fontSize: items?[index].fontSize,
+                      fontWeight: items?[index].fontWeight,
+                      fontFamily: items?[index].fontFamily,
                     ),
                   ),
                 ),
@@ -201,14 +201,14 @@ class YYDialog {
   }
 
   YYDialog listViewOfRadioButton({
-    List<RadioItem> items,
-    double height,
-    Color color,
-    Color activeColor,
-    int intialValue,
-    Function(int) onClickItemListener,
+    List<RadioItem>? items,
+    double? height,
+    Color? color,
+    Color? activeColor,
+    int? intialValue,
+    Function(int?)? onClickItemListener,
   }) {
-    Size size = MediaQuery.of(context).size;
+    Size size = MediaQuery.of(context!).size;
     return this.widget(
       Container(
         height: height,
@@ -254,7 +254,7 @@ class YYDialog {
     CustomDialog(
       gravity: gravity,
       gravityAnimationEnable: gravityAnimationEnable,
-      context: this.context,
+      context: this.context!,
       animatedFunc: animatedFunc,
       barrierColor: barrierColor,
       barrierDismissible: barrierDismissible,
@@ -297,11 +297,11 @@ class YYDialog {
               // showing or dismiss Callback
               if (isShowingChange) {
                 if (showCallBack != null) {
-                  showCallBack();
+                  showCallBack!();
                 }
               } else {
                 if (dismissCallBack != null) {
-                  dismissCallBack();
+                  dismissCallBack!();
                 }
               }
               _isShowing = isShowingChange;
@@ -313,8 +313,8 @@ class YYDialog {
   }
 
   void dismiss() {
-    if (_isShowing) {
-      Navigator.of(context, rootNavigator: true).pop();
+    if (_isShowing && context != null) {
+      Navigator.of(context!, rootNavigator: true).pop();
     }
   }
 
@@ -393,7 +393,7 @@ class YYDialog {
 ///弹窗的内容作为可变组件
 class CustomDialogChildren extends StatefulWidget {
   final List<Widget> widgetList; //弹窗内部所有组件
-  final Function(bool) isShowingChange;
+  final Function(bool)? isShowingChange;
 
   CustomDialogChildren({this.widgetList = const [], this.isShowingChange});
 
@@ -404,7 +404,9 @@ class CustomDialogChildren extends StatefulWidget {
 class CustomDialogChildState extends State<CustomDialogChildren> {
   @override
   Widget build(BuildContext context) {
-    widget.isShowingChange(true);
+    if (widget.isShowingChange != null) {
+      widget.isShowingChange!(true);
+    }
     return Column(
       children: widget.widgetList,
     );
@@ -412,7 +414,9 @@ class CustomDialogChildState extends State<CustomDialogChildren> {
 
   @override
   void dispose() {
-    widget.isShowingChange(false);
+    if (widget.isShowingChange != null) {
+      widget.isShowingChange!(false);
+    }
     super.dispose();
   }
 }
@@ -421,26 +425,26 @@ class CustomDialogChildState extends State<CustomDialogChildren> {
 class CustomDialog {
   BuildContext _context;
   Widget _child;
-  Duration _duration;
-  Color _barrierColor;
-  RouteTransitionsBuilder _transitionsBuilder;
-  bool _barrierDismissible;
-  Gravity _gravity;
-  bool _gravityAnimationEnable;
-  Function _animatedFunc;
-  EdgeInsets margin;
+  Duration? _duration;
+  Color? _barrierColor;
+  RouteTransitionsBuilder? _transitionsBuilder;
+  bool? _barrierDismissible;
+  Gravity? _gravity;
+  bool? _gravityAnimationEnable;
+  Function? _animatedFunc;
+  EdgeInsets? margin;
 
   CustomDialog({
-    @required Widget child,
-    @required BuildContext context,
-    Duration duration,
-    Color barrierColor,
-    RouteTransitionsBuilder transitionsBuilder,
-    Gravity gravity,
-    bool gravityAnimationEnable,
-    Function animatedFunc,
-    bool barrierDismissible,
-    EdgeInsets margin,
+    required Widget child,
+    required BuildContext context,
+    Duration? duration,
+    Color? barrierColor,
+    RouteTransitionsBuilder? transitionsBuilder,
+    Gravity? gravity,
+    bool? gravityAnimationEnable,
+    Function? animatedFunc,
+    bool? barrierDismissible,
+    EdgeInsets? margin,
   })  : _child = child,
         _context = context,
         _gravity = gravity,
@@ -462,7 +466,7 @@ class CustomDialog {
 
     showGeneralDialog(
       context: _context,
-      barrierColor: _barrierColor,
+      barrierColor: _barrierColor ?? Colors.white.withOpacity(0.0),
       barrierDismissible: _barrierDismissible ?? true,
       barrierLabel: "",
       transitionDuration: _duration ?? Duration(milliseconds: 250),
@@ -504,7 +508,7 @@ class CustomDialog {
       case Gravity.bottom:
       case Gravity.leftBottom:
       case Gravity.rightBottom:
-        if (margin.bottom > 0) {
+        if ((margin?.bottom ?? 0) > 0) {
           custom = Tween<Offset>(
             begin: Offset(0.0, 0.0),
             end: Offset(0.0, 0.0),
@@ -527,11 +531,11 @@ class CustomDialog {
 
     //自定义动画
     if (_animatedFunc != null) {
-      return _animatedFunc(child, animation);
+      return _animatedFunc!(child, animation);
     }
 
     //不需要默认动画
-    if (!_gravityAnimationEnable) {
+    if (!_gravityAnimationEnable!) {
       custom = Tween<Offset>(
         begin: Offset(0.0, 0.0),
         end: Offset(0.0, 0.0),
@@ -571,13 +575,13 @@ class ListTileItem {
     this.fontFamily,
   });
 
-  EdgeInsets padding;
-  Widget leading;
-  String text;
-  Color color;
-  double fontSize;
-  FontWeight fontWeight;
-  String fontFamily;
+  EdgeInsets? padding;
+  Widget? leading;
+  String? text;
+  Color? color;
+  double? fontSize;
+  FontWeight? fontWeight;
+  String? fontFamily;
 }
 
 class RadioItem {
@@ -590,11 +594,11 @@ class RadioItem {
     this.onTap,
   });
 
-  EdgeInsets padding;
-  String text;
-  Color color;
-  double fontSize;
-  FontWeight fontWeight;
-  Function(int) onTap;
+  EdgeInsets? padding;
+  String? text;
+  Color? color;
+  double? fontSize;
+  FontWeight? fontWeight;
+  Function(int)? onTap;
 }
 //============================================================================
