@@ -4,7 +4,7 @@ import 'flutter_custom_dialog.dart';
 
 class YYRadioListTile extends StatefulWidget {
   YYRadioListTile({
-    Key key,
+    Key? key,
     this.items,
     this.intialValue,
     this.color,
@@ -13,11 +13,11 @@ class YYRadioListTile extends StatefulWidget {
   })  : assert(items != null),
         super(key: key);
 
-  final List<RadioItem> items;
-  final Color color;
-  final Color activeColor;
+  final List<RadioItem>? items;
+  final Color? color;
+  final Color? activeColor;
   final intialValue;
-  final Function(int) onChanged;
+  final Function(int?)? onChanged;
 
   @override
   State<StatefulWidget> createState() {
@@ -44,19 +44,21 @@ class YYRadioListTileState extends State<YYRadioListTile> {
     return ListView.builder(
       padding: EdgeInsets.all(0.0),
       shrinkWrap: true,
-      itemCount: widget.items.length,
+      itemCount: widget.items?.length ?? 0,
       itemBuilder: (BuildContext context, int index) {
         return Material(
           color: widget.color,
           child: RadioListTile(
-            title: Text(widget.items[index].text),
+            title: Text(widget.items?[index].text ?? ""),
             value: index,
             groupValue: groupId,
             activeColor: widget.activeColor,
-            onChanged: (int value) {
+            onChanged: (int? value) {
               setState(() {
-                widget.onChanged(value);
-                groupId = value;
+                if (widget.onChanged != null) {
+                  widget.onChanged!(value);
+                }
+                groupId = value ?? -1;
               });
             },
           ),
