@@ -1,9 +1,9 @@
 class ImageEntry {
-  final String package;
+  final String? package;
 
-  final String path;
+  final String? path;
 
-  final String suffix;
+  final String? suffix;
 
   ImageEntry({this.package, this.path, this.suffix});
 }
