@@ -1,7 +1,7 @@
 class ImageResponse {
-  final String result;
-  final String suffix;
-  final String relative;
+  final String? result;
+  final String? suffix;
+  final String? relative;
 
   ImageResponse({this.result, this.suffix, this.relative});
 }
