@@ -2,8 +2,7 @@ import 'dart:collection';
 import 'dart:io';
 import 'dart:ui';
 
-import 'package:eyflutter_core/kit/utils/string/path_extension.dart';
-import 'package:eyflutter_core/kit/utils/string/string_extension.dart';
+import 'package:eyflutter_core/eyflutter_core.dart';
 import 'package:eyflutter_uikit/uikit/images/beans/image_entry.dart';
 import 'package:eyflutter_uikit/uikit/images/beans/image_response.dart';
 import 'package:eyflutter_uikit/uikit/images/beans/native_image_entry.dart';
@@ -11,8 +10,8 @@ import 'package:eyflutter_uikit/uikit/images/beans/native_image_entry.dart';
 /// flutter: imageName:{packageName:path}
 /// native: imageName:{android:name,ios:name}
 class ImageListFactory {
-  Map<String, ImageEntry> _flutterImageMap;
-  Map<String, NativeImageEntry> _nativeImageMap;
+  Map<String, ImageEntry>? _flutterImageMap;
+  Map<String, NativeImageEntry>? _nativeImageMap;
 
   ImageListFactory() {
     _flutterImageMap = new HashMap<String, ImageEntry>();
@@ -21,17 +20,17 @@ class ImageListFactory {
     nativeImageMap(_nativeImageMap);
   }
 
-  void flutterImageMap(Map<String, ImageEntry> imageMap) {}
+  void flutterImageMap(Map<String, ImageEntry>? imageMap) {}
 
-  void nativeImageMap(Map<String, NativeImageEntry> imageMap) {}
+  void nativeImageMap(Map<String, NativeImageEntry>? imageMap) {}
 
-  ImageEntry _getFlutterElementEntry(String imageName) {
+  ImageEntry? _getFlutterElementEntry(String imageName) {
     String realKey = _getRealKey(imageName);
     if (realKey.isEmptyString) {
       return null;
     }
-    var imageEntry = _flutterImageMap[realKey];
-    if (imageEntry.package.isEmptyString || imageEntry.path.isEmptyString) {
+    ImageEntry? imageEntry = _flutterImageMap?[realKey];
+    if ((imageEntry?.package).isEmptyString || (imageEntry?.path).isEmptyString) {
       return null;
     }
     return imageEntry;
@@ -46,17 +45,17 @@ class ImageListFactory {
   }
 
   String _getRealKey(String imageName) {
-    if (_flutterImageMap.containsKey(imageName)) {
+    if (_flutterImageMap?.containsKey(imageName) ?? false) {
       return imageName;
     }
     int ratio = _getRatio();
-    if (_flutterImageMap.containsKey("${imageName}_${ratio}x")) {
+    if (_flutterImageMap?.containsKey("${imageName}_${ratio}x") ?? false) {
       return "${imageName}_${ratio}x";
     }
-    if (ratio == 2 && _flutterImageMap.containsKey("${imageName}_3x")) {
+    if (ratio == 2 && (_flutterImageMap?.containsKey("${imageName}_3x") ?? false)) {
       return "${imageName}_3x";
     }
-    if (ratio == 3 && _flutterImageMap.containsKey("${imageName}_2x")) {
+    if (ratio == 3 && (_flutterImageMap?.containsKey("${imageName}_2x") ?? false)) {
       return "${imageName}_2x";
     }
     return "";
@@ -71,8 +70,8 @@ class ImageListFactory {
 
   /// 获取图片路径
   /// [imageName] 图片名
-  ImageResponse getFlutterImageResponse(String imageName) {
-    ImageEntry entry = _getFlutterElementEntry(imageName);
+  ImageResponse? getFlutterImageResponse(String imageName) {
+    ImageEntry? entry = _getFlutterElementEntry(imageName);
     if (entry == null) {
       return null;
     }
@@ -85,24 +84,24 @@ class ImageListFactory {
   /// 根据图片名称检测native图片是否存在(该方式只做过滤不保证实际存在,实际图片通过[ImageWidget]加载时作判断.)
   /// [imageAlias] 图片别名
   bool hasNativeImage(String imageAlias) {
-    return _nativeImageMap.containsKey(imageAlias);
+    return _nativeImageMap?.containsKey(imageAlias) ?? false;
   }
 
   /// 获取native图片名,通过[ImageWidget]方式加载,若返回null表示该图片不存在.
   /// [imageAlias] 图片别名
-  ImageResponse getNativeImageResponse(String imageAlias) {
-    var imageEntry = _nativeImageMap[imageAlias];
+  ImageResponse? getNativeImageResponse(String imageAlias) {
+    NativeImageEntry? imageEntry = _nativeImageMap?[imageAlias];
     if (imageEntry == null) {
       return null;
     }
-    var image = '';
+    String? image;
     var suffix = '';
     if (Platform.isIOS) {
       image = imageEntry.ios;
-      suffix = image.suffixName;
+      suffix = image?.suffixName ?? "";
     } else if (Platform.isAndroid) {
       image = imageEntry.android;
-      suffix = image.suffixName;
+      suffix = image?.suffixName ?? "";
     }
     if (image.isEmptyString) {
       return null;
@@ -124,18 +123,18 @@ class ImageListFactory {
   String getImagePathOrName(String nameOrAlias) {
     var realKey = _getRealKey(nameOrAlias);
     if (realKey.isEmptyString) {
-      NativeImageEntry entry = _nativeImageMap[nameOrAlias];
+      NativeImageEntry? entry = _nativeImageMap?[nameOrAlias];
       if (entry == null) {
         return "";
       }
       if (Platform.isAndroid) {
-        return entry.android;
+        return entry.android ?? "";
       } else if (Platform.isIOS) {
-        return entry.ios;
+        return entry.ios ?? "";
       }
       return "";
     }
-    ImageEntry entry = _flutterImageMap[realKey];
-    return "packages/${entry.package}/assets/${entry.path}";
+    ImageEntry? entry = _flutterImageMap?[realKey];
+    return "packages/${entry?.package}/assets/${entry?.path}";
   }
 }
