@@ -8,8 +8,8 @@ import 'package:flutter/material.dart';
 
 class Mp4ImageProvider extends ImageProvider<Mp4ImageProvider> {
   /// 图片名称
-  final String imageName;
-  final String suffix;
+  final String? imageName;
+  final String? suffix;
   final double scale;
 
   String mp4ImageProviderMethodName = "3cf4e84a8e61d381";
@@ -21,14 +21,14 @@ class Mp4ImageProvider extends ImageProvider<Mp4ImageProvider> {
     return MultiFrameImageStreamCompleter(codec: _loadAsync(key), scale: key.scale);
   }
 
-  Future<T> getNativeImage<T>(String imageName, String suffix) {
+  Future<T?>? getNativeImage<T>(String imageName, String suffix) {
     return CloudChannelManager.instance.channel
-        .invokeMethod<T>(mp4ImageProviderMethodName, {"imageName": imageName, "suffix": suffix});
+        ?.invokeMethod<T>(mp4ImageProviderMethodName, {"imageName": imageName, "suffix": suffix});
   }
 
   Future<Codec> _loadAsync(Mp4ImageProvider key) async {
-    Uint8List bytes = await getNativeImage<Uint8List>(imageName, suffix);
-    return instantiateImageCodec(bytes);
+    Uint8List? bytes = await getNativeImage<Uint8List>(imageName ?? "", suffix ?? "");
+    return instantiateImageCodec(bytes!);
   }
 
   @override
