@@ -3,18 +3,18 @@ import 'package:flutter/material.dart';
 
 class CustomLoadListView extends StatefulWidget {
   /// 空视图
-  final Widget emptyView;
+  final Widget? emptyView;
 
   /// 列表数据
-  final List list;
+  final List? list;
 
   /// 条目构建
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
 
   /// 条目分隔线
-  final IndexedWidgetBuilder separatorBuilder;
+  final IndexedWidgetBuilder? separatorBuilder;
 
-  const CustomLoadListView({Key key, this.emptyView, this.list, this.itemBuilder, this.separatorBuilder})
+  const CustomLoadListView({Key? key, this.emptyView, this.list, this.itemBuilder, this.separatorBuilder})
       : super(key: key);
 
   @override
@@ -47,8 +47,8 @@ class _CustomLoadListViewState extends State<CustomLoadListView> {
       offstage: widget.list.isEmptyList,
       child: ListView.separated(
           controller: _scrollController,
-          itemBuilder: widget.itemBuilder,
-          separatorBuilder: widget.separatorBuilder,
+          itemBuilder: widget.itemBuilder!,
+          separatorBuilder: widget.separatorBuilder!,
           itemCount: (widget.list?.length ?? 0 + 1)),
     );
   }
