@@ -12,15 +12,15 @@ class RefreshFooterController extends StatefulWidget {
       this.loadTriggerPullDistance,
       this.loadIndicatorWidget});
 
-  final double loadIndicatorExtent;
-  final double pulledExtent;
-  final bool noMore;
-  final bool enableInfiniteLoad;
-  final LoadMode mode;
-  final double loadTriggerPullDistance;
+  final double? loadIndicatorExtent;
+  final double? pulledExtent;
+  final bool? noMore;
+  final bool? enableInfiniteLoad;
+  final LoadMode? mode;
+  final double? loadTriggerPullDistance;
 
   //加载指示效果组件
-  final Widget loadIndicatorWidget;
+  final Widget? loadIndicatorWidget;
 
   @override
   _RefreshFooterControllerState createState() => _RefreshFooterControllerState();
@@ -44,9 +44,12 @@ class _RefreshFooterControllerState extends State<RefreshFooterController> {
   @override
   Widget build(BuildContext context) {
     // 是否到达触发加载距离
-    overTriggerDistance = widget.mode != LoadMode.inactive && widget.pulledExtent >= widget.loadTriggerPullDistance;
+    overTriggerDistance =
+        widget.mode != LoadMode.inactive && ((widget.pulledExtent ?? 0) >= (widget.loadTriggerPullDistance ?? 0));
     _isShowIndicator = (widget.mode == LoadMode.done || widget.mode == LoadMode.loaded) ? false : true;
-    _footerHeight = widget.loadIndicatorExtent > widget.pulledExtent ? widget.loadIndicatorExtent : widget.pulledExtent;
+    _footerHeight = (widget.loadIndicatorExtent ?? 0) > (widget.pulledExtent ?? 0)
+        ? (widget.loadIndicatorExtent ?? 0)
+        : (widget.pulledExtent ?? 0);
     return Container(
       height: _footerHeight,
       child: buildDefaultFooter(),
@@ -74,11 +77,11 @@ class _RefreshFooterControllerState extends State<RefreshFooterController> {
         Offstage(
           offstage: _isShowIndicator && widget.loadIndicatorWidget != null,
           child: Text(
-            _showText ?? "",
+            _showText,
             textAlign: TextAlign.center,
             style: TextStyle(
                 letterSpacing: 0.82,
-                color: Color(RefreshInitManager.instance.loadHandler().loadTextColor() ?? 0xff666666),
+                color: Color(RefreshInitManager.instance.loadHandler()?.loadTextColor() ?? 0xffffffff),
                 decoration: TextDecoration.none,
                 fontSize: 14),
           ),
@@ -89,7 +92,7 @@ class _RefreshFooterControllerState extends State<RefreshFooterController> {
 
   String get _showText {
     if (widget.noMore ?? false) {
-      return RefreshInitManager.instance.loadHandler().noDataText();
+      return RefreshInitManager.instance.loadHandler()?.noDataText() ?? "";
     }
     // if (widget.enableInfiniteLoad) {
     //   if (widget.mode == LoadMode.done || widget.mode == LoadMode.loaded) {
@@ -102,15 +105,15 @@ class _RefreshFooterControllerState extends State<RefreshFooterController> {
     switch (widget.mode) {
       case LoadMode.load:
       case LoadMode.armed:
-        return RefreshInitManager.instance.loadHandler().loadingText();
+        return RefreshInitManager.instance.loadHandler()?.loadingText() ?? "";
       case LoadMode.loaded:
       case LoadMode.done:
-        return RefreshInitManager.instance.loadHandler().loadCompleteText();
+        return RefreshInitManager.instance.loadHandler()?.loadCompleteText() ?? "";
       default:
         if (overTriggerDistance) {
-          return RefreshInitManager.instance.loadHandler().loadReleaseText();
+          return RefreshInitManager.instance.loadHandler()?.loadReleaseText() ?? "";
         } else {
-          return RefreshInitManager.instance.loadHandler().loadIdleText();
+          return RefreshInitManager.instance.loadHandler()?.loadIdleText() ?? "";
         }
     }
   }
