@@ -6,52 +6,48 @@ import 'cs_ring_painter.dart';
 
 /// 自转环效果
 class SpinRing extends StatefulWidget {
+  final Color? indicatorBackgroundColor;
+  final Color indicatorColor;
+  final double size;
+  final double lineWidth;
+  final Duration duration;
+  final AnimationController? controller;
+
   const SpinRing({
-    Key key,
-    @required this.indicatorColor,
+    Key? key,
+    required this.indicatorColor,
     this.indicatorBackgroundColor,
     this.lineWidth = 7.0,
     this.size = 50.0,
     this.duration = const Duration(milliseconds: 1200),
     this.controller,
-  })  : assert(indicatorColor != null),
-        assert(lineWidth != null),
-        assert(size != null),
-        super(key: key);
-
-  final Color indicatorBackgroundColor;
-  final Color indicatorColor;
-  final double size;
-  final double lineWidth;
-  final Duration duration;
-  final AnimationController controller;
+  });
 
   @override
   _SpinKitRingState createState() => _SpinKitRingState();
 }
 
 class _SpinKitRingState extends State<SpinRing> with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _animation1, _animation2, _animation3;
+  AnimationController? _controller;
+  Animation<double>? _animation1, _animation2, _animation3;
 
   @override
   void initState() {
     super.initState();
-
     _controller = (widget.controller ?? AnimationController(vsync: this, duration: widget.duration))
       ..addListener(() => setState(() {}))
       ..repeat();
     _animation1 = Tween(begin: 0.0, end: 1.0)
-        .animate(CurvedAnimation(parent: _controller, curve: const Interval(0.0, 1.0, curve: Curves.linear)));
+        .animate(CurvedAnimation(parent: _controller!, curve: const Interval(0.0, 1.0, curve: Curves.linear)));
     _animation2 = Tween(begin: -2 / 3, end: 1 / 2)
-        .animate(CurvedAnimation(parent: _controller, curve: const Interval(0.5, 1.0, curve: Curves.linear)));
+        .animate(CurvedAnimation(parent: _controller!, curve: const Interval(0.5, 1.0, curve: Curves.linear)));
     _animation3 = Tween(begin: 0.25, end: 5 / 6)
-        .animate(CurvedAnimation(parent: _controller, curve: const Interval(0.0, 1.0, curve: SpinRingCurve())));
+        .animate(CurvedAnimation(parent: _controller!, curve: const Interval(0.0, 1.0, curve: SpinRingCurve())));
   }
 
   @override
   void dispose() {
-    _controller.dispose();
+    _controller?.dispose();
     super.dispose();
   }
 
@@ -59,7 +55,7 @@ class _SpinKitRingState extends State<SpinRing> with SingleTickerProviderStateMi
   Widget build(BuildContext context) {
     return Center(
       child: Transform(
-        transform: Matrix4.identity()..rotateZ((_animation1.value) * 5 * pi / 6),
+        transform: Matrix4.identity()..rotateZ((_animation1?.value ?? 0) * 5 * pi / 6),
         alignment: FractionalOffset.center,
         child: SizedBox.fromSize(
           size: Size.square(widget.size),
@@ -67,8 +63,8 @@ class _SpinKitRingState extends State<SpinRing> with SingleTickerProviderStateMi
             foregroundPainter: CsRingPainter(
               paintWidth: widget.lineWidth,
               trackColor: widget.indicatorBackgroundColor ?? Colors.transparent,
-              progressPercent: _animation3.value,
-              startAngle: pi * _animation2.value,
+              progressPercent: (_animation3?.value ?? 0),
+              startAngle: pi * (_animation2?.value ?? 0),
             ),
             painter: CsRingPainter(
               paintWidth: widget.lineWidth,
