@@ -1,22 +1,18 @@
-import 'package:eyflutter_core/lang/lang_manager.dart';
-import 'package:eyflutter_core/mq/ebus/cloud_ebus.dart';
+import 'package:eyflutter_core/eyflutter_core.dart';
 import 'package:eyflutter_uikit/uikit/state/lifecycle_state.dart';
 import 'package:flutter/material.dart';
 
-abstract class BaseState<T extends StatefulWidget> extends LifecycleState<T> {
-  Locale _locale;
-
-  /// 更新语言监听
-  void onChangedLanguage() {}
+abstract class BaseState<T extends StatefulWidget> extends LifecycleState<T> with OnLangChangeState {
+  Locale? _locale;
 
   @override
   void onCreate(BuildContext context, String routeName, Map<String, dynamic> arguments) {
     _locale = Localizations.localeOf(context);
-    CloudEBus.instance.addEvent(this, action: LangManager.instance.stateAction, method: _changeLocale);
+    LangManager.instance.addChangeState(this.hashCode, this);
   }
 
-  _changeLocale(Locale locale) {
-    onChangedLanguage();
+  @override
+  void dispatchLangState(Locale locale) {
     setState(() {
       _locale = locale;
     });
@@ -24,11 +20,11 @@ abstract class BaseState<T extends StatefulWidget> extends LifecycleState<T> {
 
   @override
   void onDestroy(String routeName, Map<String, dynamic> arguments) {
-    CloudEBus.instance.disposeEvent(this, action: LangManager.instance.stateAction);
+    LangManager.instance.removeChangeState(this.hashCode);
   }
 
   ///在@override#build中调用
-  Widget overrideLocalizations(BuildContext context, {Widget child}) {
+  Widget overrideLocalizations({required BuildContext context, required Widget child}) {
     return Localizations.override(
       context: context,
       locale: _locale,
