@@ -21,25 +21,25 @@ mixin RoundListening {
 
 class RoundWidget extends StatefulWidget {
   /// 视图宽高比
-  final double aspectRatio;
+  final double? aspectRatio;
 
   /// 子组件集合
-  final List<Widget> children;
+  final List<Widget>? children;
 
   /// 是否自动播放
-  final bool autoStart;
+  final bool? autoStart;
 
   /// 部件构建完成回调
-  final RoundWidgetBuildCallback buildCall;
+  final RoundWidgetBuildCallback? buildCall;
 
   /// 视图切换回调
-  final RoundChangedCallback changedCall;
+  final RoundChangedCallback? changedCall;
 
   /// Swiper page indicator
-  final SwiperIndicator indicator;
+  final SwiperIndicator? indicator;
 
   const RoundWidget(
-      {Key key, this.aspectRatio, this.children, this.autoStart, this.buildCall, this.changedCall, this.indicator})
+      {Key? key, this.aspectRatio, this.children, this.autoStart, this.buildCall, this.changedCall, this.indicator})
       : super(key: key);
 
   @override
@@ -47,36 +47,36 @@ class RoundWidget extends StatefulWidget {
 }
 
 class _RoundWidgetState extends State<RoundWidget> with RoundListening {
-  SwiperController _controller;
+  SwiperController? _controller;
 
   @override
   void initState() {
     _controller = SwiperController();
-    _controller.addListener(() {
+    _controller!.addListener(() {
       if (widget.changedCall != null) {
-        widget.changedCall(_controller.index);
+        widget.changedCall!(_controller!.index);
       }
     });
     if (widget.buildCall != null) {
-      widget.buildCall(_controller, this);
+      widget.buildCall!(_controller!, this);
     }
     super.initState();
   }
 
   @override
   void dispose() {
-    _controller.dispose();
+    _controller!.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return AspectRatio(
-      aspectRatio: widget.aspectRatio,
+      aspectRatio: widget.aspectRatio ?? 0,
       child: Swiper(
         children: widget.children,
         controller: _controller,
-        autoStart: widget.autoStart,
+        autoStart: widget.autoStart ?? true,
         circular: true,
         indicator: widget.indicator,
       ),
@@ -85,21 +85,21 @@ class _RoundWidgetState extends State<RoundWidget> with RoundListening {
 
   @override
   void next() {
-    _controller.nextPage(duration: Duration(milliseconds: 200), curve: Curves.easeOut);
+    _controller?.nextPage(duration: Duration(milliseconds: 200), curve: Curves.easeOut);
   }
 
   @override
   void previous() {
-    _controller.previousPage(duration: Duration(milliseconds: 200), curve: Curves.easeOut);
+    _controller?.previousPage(duration: Duration(milliseconds: 200), curve: Curves.easeOut);
   }
 
   @override
   void start() {
-    _controller.start();
+    _controller?.start();
   }
 
   @override
   void stop() {
-    _controller.stop();
+    _controller?.stop();
   }
 }
