@@ -7,16 +7,16 @@ class ItemViewBuilder {
   ItemViewBuilder({this.height, this.backgroundColor, this.children = const <Widget>[], this.rowPadding});
 
   /// 条目视图高度
-  final double height;
+  final double? height;
 
   /// 条目背景颜色
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// 子视图
   final List<Widget> children;
 
   /// 行边距
-  final EdgeInsetsGeometry rowPadding;
+  final EdgeInsetsGeometry? rowPadding;
 }
 
 /// 横向滑动视图
@@ -33,30 +33,30 @@ class HorScrollView extends StatelessWidget {
       this.rowPadding});
 
   /// 头部视图
-  final Widget topContainer;
+  final Widget? topContainer;
 
   final List<ItemViewBuilder> children;
 
   /// 底部视图
-  final Widget bottomContainer;
+  final Widget? bottomContainer;
 
   /// 分隔线宽度
-  final double dividerWidth;
+  final double? dividerWidth;
 
   /// 分隔线颜色
-  final Color dividerColor;
+  final Color? dividerColor;
 
   /// 横向分隔线高度
-  final double horDividerHeight;
+  final double? horDividerHeight;
 
   /// 横向分隔线颜色
-  final Color horDividerColor;
+  final Color? horDividerColor;
 
   /// 背景颜色
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// 行边距
-  final EdgeInsetsGeometry rowPadding;
+  final EdgeInsetsGeometry? rowPadding;
 
   @override
   Widget build(BuildContext context) {
@@ -69,10 +69,11 @@ class HorScrollView extends StatelessWidget {
   Column _buildViews(List<ItemViewBuilder> children) {
     List<Widget> rows = [this.topContainer ?? Container()];
     LinkedHashSet<Widget> widgets = new LinkedHashSet();
-    var count = children?.length;
-    children?.forEach((ItemViewBuilder builder) {
+    var count = children.length;
+    children.forEach((ItemViewBuilder builder) {
       var index = children.indexOf(builder);
-      widgets.add(_buildRowView(backgroundColor, rowPadding, builder));
+      widgets.add(_buildRowView(backgroundColor ?? Colors.white,
+          rowPadding ?? EdgeInsets.only(left: 0, top: 0, right: 0, bottom: 0), builder));
       if (((index + 1) < count)) {
         widgets.add(Container(
           height: horDividerHeight ?? 0,
@@ -90,14 +91,14 @@ class HorScrollView extends StatelessWidget {
 
   Container _buildRowView(Color backgroundColor, EdgeInsetsGeometry rowPadding, ItemViewBuilder builder) {
     return Container(
-      height: builder?.height ?? 0,
-      color: (builder?.backgroundColor ?? backgroundColor) ?? Colors.white,
-      padding: (builder?.rowPadding ?? rowPadding) ?? EdgeInsets.only(left: 0, top: 0, right: 0, bottom: 0),
+      height: builder.height ?? 0,
+      color: (builder.backgroundColor ?? backgroundColor),
+      padding: (builder.rowPadding ?? rowPadding),
       child: ScrollConfiguration(
         behavior: OverScrollBehavior(),
         child: ListView.separated(
           scrollDirection: Axis.horizontal,
-          itemCount: builder?.children?.length ?? 0,
+          itemCount: builder.children.length,
           separatorBuilder: (context, index) {
             return Container(
               width: this.dividerWidth ?? 0,
