@@ -2,22 +2,16 @@ import 'package:flutter/material.dart';
 
 class IconWidget extends Icon {
   IconWidget(IconData icon,
-      {Key key,
-      double size,
-      Color color,
-      String semanticLabel,
-      TextDirection textDirection,
+      {Key? key,
+      double? size,
+      Color? color,
+      String? semanticLabel,
+      TextDirection? textDirection,
       this.isVisible = true,
-      this.padding =
-          const EdgeInsets.only(left: 0, top: 0, right: 0, bottom: 0),
+      this.padding = const EdgeInsets.only(left: 0, top: 0, right: 0, bottom: 0),
       this.width,
       this.height})
-      : super(icon,
-            key: key,
-            size: size,
-            color: color,
-            semanticLabel: semanticLabel,
-            textDirection: textDirection);
+      : super(icon, key: key, size: size, color: color, semanticLabel: semanticLabel, textDirection: textDirection);
 
   //是否显示组件，默认显示
   final bool isVisible;
@@ -26,10 +20,10 @@ class IconWidget extends Icon {
   final EdgeInsetsGeometry padding;
 
   /// 控件宽度
-  final double width;
+  final double? width;
 
   /// 控件高度
-  final double height;
+  final double? height;
 
   @override
   Widget build(BuildContext context) {
